Remove await from writeRowGroup calls [bjf/jb]
So this is a commit that may have to be taken on faith.  In testing, we
found that when we wrote data with sufficiently large schema that
exceeded the default row group size, the `await` on writeRowGroup caused
the call to be handled asynchronously, meaning that rowBuffer would not
be reset to an empty object, which means that we would append an
ever-increasing number of rows for each invocation of `appendRow()`.
This caused files that should have 4100-ish rows to instead have over
100,000.  By removing the await, the library now handles the
`writeRowGroup` call synchronously, and all files, irrespective of row
count, contain expected data.

Attempts to write an integration test within the current framework have
proved difficult, but at least this commit doesn't break existing tests.